requirements.txt
# ResumePard - Advanced Resume-Job Description Matching System
# Core Dependencies

# Natural Language Processing
spacy>=3.8.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.2.0

# Data Processing
numpy>=2.0.0
pandas>=2.0.0

# Text Processing
lxml>=6.0.0

# Web Application
streamlit>=1.28.0
plotly>=5.17.0

# Language Detection and Processing
langcodes>=3.5.0
markdown-it-py>=3.0.0

# Additional NLP Support
marisa-trie>=1.2.0

# Command Line Interface
click>=8.1.0

# Development Dependencies (Optional)
# Uncomment for development environment
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.910

# Note: After installation, run:
# python -m spacy download en_core_web_sm
# Web Interface  
streamlit>=1.28.0  
pandas>=2.0.0 
